third_party/WebKit/LayoutTests/fast/ime/inputmethodcontext-composition-expected.txt

This tests composition attribute of InputMethodContext.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS composition.text is ""
PASS composition2.text is ""
Testing IME on an input element.
PASS composition.text is "abcde"
PASS composition2.text is ""
PASS composition.selectionStart is 1
PASS composition.selectionEnd is 4
PASS composition.getSegments().length is 3
PASS composition.getSegments().length is 2
PASS composition.getSegments().length is 2
PASS composition.text is ""
PASS composition.getSegments().length is 0
PASS input.value is "abcde"
PASS input2.value is ""
Testing IME on another input element.
PASS input.value is "abcde"
PASS input2.value is ""
PASS composition.getSegments().length is 0
PASS composition2.getSegments().length is 1
PASS input.value is "abcde"
PASS input2.value is "fghij"
PASS composition.text is ""
PASS composition2.text is "fghij"
PASS composition2.text is ""
PASS input.value is "abcde"
PASS input2.value is "fghij"
Testing when context is obtained after some IME input.
PASS input3.value is ""
PASS input3.value is "klmno"
PASS composition3.text is "klmno"
PASS composition3.text is ""
PASS successfullyParsed is true

TEST COMPLETE
